+++ generators/app/tasks/appSetup/files.js
@@ -60,8 +60,6 @@ module.exports.CUSTOM_TEXT_PATH = `${module.exports.COMPONENTS_PATH}/CustomText`
 module.exports.CUSTOM_TEXT_INPUT_PATH = `${module.exports.COMPONENTS_PATH}/CustomTextInput`;
 // APP COMPONENTS - LOADABLE
 module.exports.LOADABLE_PATH = `${module.exports.COMPONENTS_PATH}/Loadable`;
-// APP COMPONENTS - WITH FORMIK FIELD
-module.exports.WITH_FORMIK_FIELD_PATH = `${module.exports.COMPONENTS_PATH}/withFormikField`;
 
 // APP SCREENS - HOME
 module.exports.HOME_PATH = `${module.exports.SCREENS_PATH}/Home`;