+++ generators/app/templates/__mocks__/setup.js
@@ -1,1 +1,3 @@
 jest.mock('react-native/Libraries/Animated/src/NativeAnimatedHelper');
+global.window = {};
+global.window = global;