+++ generators/app/templates/src/app/screens/Auth/screens/SignUp/index.tsx
@@ -1,18 +1,17 @@
-import React, { useCallback } from 'react';
+import React from 'react';
 import { Keyboard, TouchableWithoutFeedback, ScrollView, View } from 'react-native';
 import KeyboardSpacer from 'react-native-keyboard-spacer';
 import i18next from 'i18next';
-import { Formik } from 'formik';
+import { useForm } from 'react-hook-form';
 import CustomButton from '@components/CustomButton';
 import CustomText from '@components/CustomText';
-import { CustomTextInputFormikField } from '@components/CustomTextInput';
+import ControlledCustomTextInput from '@components/CustomTextInput/controller';
 import { isIos } from '@constants/platform';
 import { useAsyncRequest } from '@hooks/useRequest';
 import { Navigation } from '@interfaces/navigation';
-import { FIELDS, SIGNUP_INITIAL_VALUES } from '@screens/Auth/constants';
+import { FIELDS, SignupFormValues } from '@screens/Auth/constants';
 import * as AuthService from '@services/AuthService';
 import {
-  validationsWrapper,
   validateRequired,
   validateEmail,
   validateOnlyText,
@@ -27,101 +26,117 @@ function SignUp({ navigation }: Navigation) {
     request: AuthService.signup,
     withPostSuccess: () => navigation.goBack()
   });
+
+  const {
+    control,
+    errors,
+    formState: { isValid },
+    handleSubmit
+  } = useForm<SignupFormValues>({ mode: 'onTouched' });
+
   const hasSignUpError = !!error;
-  const handleSignUp = useCallback(
-    values => {
-      Keyboard.dismiss();
-      signUp(values);
-    },
-    [signUp]
-  );
+  const handleSignUp = (values: SignupFormValues) => {
+    Keyboard.dismiss();
+    signUp(values);
+  };
+
   return (
-    <Formik onSubmit={handleSignUp} initialValues={SIGNUP_INITIAL_VALUES}>
-      {({ handleSubmit, isValid }) => (
-        <>
-          <ScrollView
-            bounces={false}
-            keyboardShouldPersistTaps="handled"
-            showsVerticalScrollIndicator={false}
-            contentContainerStyle={styles.container}
-            style={styles.stretchAndFlex}>
-            <TouchableWithoutFeedback onPress={Keyboard.dismiss}>
-              <View style={[styles.stretchAndFlex, styles.form]}>
-                <CustomTextInputFormikField
-                  animated
-                  label={i18next.t('SIGNUP:NAME')}
-                  name={FIELDS.name}
-                  showError={hasSignUpError}
-                  validate={validationsWrapper([validateRequired, validateOnlyText])}
-                />
-                <CustomTextInputFormikField
-                  animated
-                  label={i18next.t('SIGNUP:SURNAME')}
-                  name={FIELDS.surname}
-                  showError={hasSignUpError}
-                  validate={validationsWrapper([validateRequired, validateOnlyText])}
-                />
-                <CustomTextInputFormikField
-                  animated
-                  label={i18next.t('SIGNUP:BIRTH_DATE')}
-                  name={FIELDS.birthDate}
-                  placeholder={i18next.t('SIGNUP:BIRTH_DATE_PLACEHOLDER')}
-                  showError={hasSignUpError}
-                  validate={validateRequired}
-                />
-                <CustomTextInputFormikField
-                  animated
-                  label={i18next.t('SIGNUP:SEX')}
-                  name={FIELDS.sex}
-                  placeholder={i18next.t('SIGNUP:SEX_PLACEHOLDER')}
-                  showError={hasSignUpError}
-                  validate={validationsWrapper([validateRequired, validateOnlyText])}
-                />
-                <CustomTextInputFormikField
-                  animated
-                  keyboardType="email-address"
-                  label={i18next.t('SIGNUP:MAIL')}
-                  name={FIELDS.email}
-                  placeholder={i18next.t('SIGNUP:MAIL_PLACEHOLDER')}
-                  showError={hasSignUpError}
-                  validate={validationsWrapper([validateRequired, validateEmail])}
-                />
-                <CustomTextInputFormikField
-                  animated
-                  showEye
-                  secureTextEntry
-                  label={i18next.t('SIGNUP:PASSWORD')}
-                  name={FIELDS.password}
-                  showError={hasSignUpError}
-                  validate={validationsWrapper([validateRequired, validateMinLength(8)])}
-                />
-                <CustomTextInputFormikField
-                  animated
-                  isOptional
-                  keyboardType="phone-pad"
-                  label={i18next.t('SIGNUP:PHONE_NUMBER')}
-                  name={FIELDS.phoneNumber}
-                  placeholder={i18next.t('SIGNUP:PHONE_NUMBER_PLACEHOLDER')}
-                  showError={hasSignUpError}
-                />
-                {hasSignUpError && (
-                  <CustomText error center>
-                    {i18next.t('SIGNUP:SIGNUP_FAILURE')}
-                  </CustomText>
-                )}
-                <CustomButton
-                  onPress={handleSubmit}
-                  style={styles.formButton}
-                  title={i18next.t('SIGNUP:SIGN_UP')}
-                  disabled={hasSignUpError || !isValid}
-                />
-              </View>
-            </TouchableWithoutFeedback>
-          </ScrollView>
-          {isIos && <KeyboardSpacer />}
-        </>
-      )}
-    </Formik>
+    <>
+      <ScrollView
+        bounces={false}
+        keyboardShouldPersistTaps="handled"
+        showsVerticalScrollIndicator={false}
+        contentContainerStyle={styles.container}
+        style={styles.stretchAndFlex}>
+        <TouchableWithoutFeedback onPress={Keyboard.dismiss}>
+          <View style={[styles.stretchAndFlex, styles.form]}>
+            <ControlledCustomTextInput
+              control={control}
+              animated
+              label={i18next.t('SIGNUP:NAME')}
+              name={FIELDS.name}
+              showError={hasSignUpError}
+              error={errors[FIELDS.name]?.message}
+              rules={{ ...validateRequired, ...validateOnlyText }}
+            />
+            <ControlledCustomTextInput
+              control={control}
+              animated
+              label={i18next.t('SIGNUP:SURNAME')}
+              name={FIELDS.surname}
+              showError={hasSignUpError}
+              error={errors[FIELDS.surname]?.message}
+              rules={{ ...validateRequired, ...validateOnlyText }}
+            />
+            <ControlledCustomTextInput
+              control={control}
+              animated
+              label={i18next.t('SIGNUP:BIRTH_DATE')}
+              name={FIELDS.birthDate}
+              placeholder={i18next.t('SIGNUP:BIRTH_DATE_PLACEHOLDER')}
+              showError={hasSignUpError}
+              error={errors[FIELDS.birthDate]?.message}
+              rules={validateRequired}
+            />
+            <ControlledCustomTextInput
+              control={control}
+              animated
+              label={i18next.t('SIGNUP:SEX')}
+              name={FIELDS.sex}
+              placeholder={i18next.t('SIGNUP:SEX_PLACEHOLDER')}
+              showError={hasSignUpError}
+              error={errors[FIELDS.sex]?.message}
+              rules={{ ...validateRequired, ...validateOnlyText }}
+            />
+            <ControlledCustomTextInput
+              control={control}
+              animated
+              keyboardType="email-address"
+              label={i18next.t('SIGNUP:MAIL')}
+              name={FIELDS.email}
+              placeholder={i18next.t('SIGNUP:MAIL_PLACEHOLDER')}
+              showError={hasSignUpError}
+              error={errors[FIELDS.email]?.message}
+              rules={{ ...validateRequired, ...validateEmail }}
+            />
+            <ControlledCustomTextInput
+              control={control}
+              animated
+              showEye
+              secureTextEntry
+              label={i18next.t('SIGNUP:PASSWORD')}
+              name={FIELDS.password}
+              showError={hasSignUpError}
+              error={errors[FIELDS.password]?.message}
+              rules={{ ...validateRequired, ...validateMinLength(8) }}
+            />
+            <ControlledCustomTextInput
+              control={control}
+              animated
+              isOptional
+              keyboardType="phone-pad"
+              label={i18next.t('SIGNUP:PHONE_NUMBER')}
+              name={FIELDS.phoneNumber}
+              placeholder={i18next.t('SIGNUP:PHONE_NUMBER_PLACEHOLDER')}
+              error={errors[FIELDS.phoneNumber]?.message}
+              showError={hasSignUpError}
+            />
+            {hasSignUpError && (
+              <CustomText error center>
+                {i18next.t('SIGNUP:SIGNUP_FAILURE')}
+              </CustomText>
+            )}
+            <CustomButton
+              onPress={handleSubmit(handleSignUp)}
+              style={styles.formButton}
+              title={i18next.t('SIGNUP:SIGN_UP')}
+              disabled={hasSignUpError || !isValid}
+            />
+          </View>
+        </TouchableWithoutFeedback>
+      </ScrollView>
+      {isIos && <KeyboardSpacer />}
+    </>
   );
 }